api/requirements.txt
uvicorn==0.35.0
fastapi==0.116.1
PyMySQL==1.1.1
sqlmodel==0.0.24
pwdlib[argon2]==0.2.1
pyjwt==2.10.1
pydantic_settings==2.10.1
pydantic[email]==2.11.7
httpx==0.28.1